Make the shard size configurable - 2

+++ recover/lrc.go
@@ -136,7 +136,7 @@ func (lrch *LRCHandler) RecoverShardStage(shdinfo *lrcpkg.Shardsinfo, td message
 		peer := td.Locations[idx]
 		for r := 1; r < 6; r++ {
 			shard, err = lrch.le.GetShard(peer.NodeId, base58.Encode(td.Id), peer.Addrs, td.Hashs[idx], num, sw, tasklife)
-			if err == nil && len(shard) == lrch.le.sn.YTFS().Meta().DataBlockSize {
+			if err == nil && len(shard) == (int)(lrch.le.sn.YTFS().Meta().DataBlockSize) {
 				break
 			}
 
@@ -147,7 +147,7 @@ func (lrch *LRCHandler) RecoverShardStage(shdinfo *lrcpkg.Shardsinfo, td message
 			<-time.After(time.Millisecond * 50)
 		}
 
-		if len(shard) != lrch.le.sn.YTFS().Meta().DataBlockSize  {
+		if len(shard) != (int)(lrch.le.sn.YTFS().Meta().DataBlockSize)  {
 			log.Println("[recover] error: shard lenth != 16K, missidx=", idx)
 			continue
 		}
@@ -265,7 +265,7 @@ effortwk:
 			}
 		}
 
-		if len(shard) < lrch.le.sn.YTFS().Meta().DataBlockSize  {
+		if len(shard) < (int)(lrch.le.sn.YTFS().Meta().DataBlockSize)  {
 			log.Println("[recover][ytlrc] shard is empty or get error!! idx=", idx)
 			indexs2 = append(indexs2, idx)
 			continue
